start common fnc file
