+++ packages/augur-simplified/src/modules/sidebar/sidebar.tsx
@@ -21,21 +21,7 @@ import {
   marketStatusItems,
   sortByItems,
 } from '../constants';
-
-export function isMobileBrowser(): boolean {
-  return (
-    navigator.userAgent.indexOf('Android') > -1 ||
-    navigator.userAgent.indexOf('webOS') > -1 ||
-    navigator.userAgent.indexOf('Windows Phone') > -1
-  );
-}
-
-export function isMobileBrowserTall(): boolean {
-  return (
-    navigator.userAgent.indexOf('iPhone') > -1 ||
-    navigator.userAgent.indexOf('iPad') > -1
-  );
-}
+import { isMobileBrowser, isMobileBrowserTall } from '../../utils/common-functions';
 
 interface SideBarHeaderProps {
   header?: string;
